pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "ShahProgrammer", email = "shahprogrammer05@gmail.com"},
]
dependencies = [
    "rich>=12.6.0",
    "pycryptodomex>=3.16.0",
    "psycopg[binary]>=3.1.4",
    "proto-plus>=1.22.1",
    "pyperclip>=1.8.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
